+++ tiledb/sm/cpp_api/vfs.h
@@ -452,6 +452,26 @@ class VFS {
     ctx.handle_error(tiledb_vfs_remove_file(ctx, vfs_.get(), uri.c_str()));
   }
 
+  /** Retrieves the size of a directory with the input URI. */
+  uint64_t dir_size(const std::string& uri) const {
+    uint64_t ret;
+    auto& ctx = ctx_.get();
+    ctx.handle_error(tiledb_vfs_dir_size(ctx, vfs_.get(), uri.c_str(), &ret));
+    return ret;
+  }
+
+  /**
+   * Retrieves the children in directory `uri`. This function is
+   * non-recursive, i.e., it focuses in one level below `uri`.
+   */
+  std::vector<std::string> ls(const std::string& uri) const {
+    std::vector<std::string> ret;
+    auto& ctx = ctx_.get();
+    ctx.handle_error(
+        tiledb_vfs_ls(ctx, vfs_.get(), uri.c_str(), ls_getter, &ret));
+    return ret;
+  }
+
   /** Retrieves the size of a file with the input URI. */
   uint64_t file_size(const std::string& uri) const {
     uint64_t ret;
@@ -493,6 +513,25 @@ class VFS {
   /** Get the config **/
   Config config() const {
     return config_;
+  }
+
+  /* ********************************* */
+  /*          STATIC FUNCTIONS         */
+  /* ********************************* */
+
+  /**
+   * Callback function to be used when invoking the C TileDB function
+   * for getting the children of a URI. It simply adds `path` to
+   * `vec` (which is casted from `data`).
+   *
+   * @param path The path of a visited TileDB object
+   * @param data This will be casted to the vector that will store `path`.
+   * @return If `1` then the walk should continue to the next object.
+   */
+  static int ls_getter(const char* path, void* data) {
+    auto vec = static_cast<std::vector<std::string>*>(data);
+    vec->emplace_back(path);
+    return 1;
   }
 
  private: